# Use app locks for all transactions