@Charles-Gagnon 2022-11-07
> What does this do for perf when there's only one function running at a time? Is calling `sp_getapplock` an expensive operation?

Added results from single-host perf tests to description (there didn't seem to be any noticeable impact)