@chlafreniere 2022-11-07
Thanks for setting up the benchmarks with/without app locking. Really glad that the runtime impact time is very minor.