@JatinSanghvi 2022-11-07
Should we add app lock in queries inside `SqlTriggerListener.cs` too? That might fix https://github.com/Azure/azure-functions-sql-extension/issues/406.